Refactor Isolate communication
As before, the communication between the MainIsolate
and the ForegroundServiceIsolate is done via a method
forwarding system: each class is instanciated on both
side and the messaging system forwards methods calls
on one side to method calls on the other side.

Refactor GpsStatusNotifier by extracting ForwardingGpsStatusNotifier.
Refactor slightly the Isolate code for better readability and
consistency.